+++ rmgpy/data/kinetics/common.py
@@ -33,27 +33,11 @@ This module contains classes and functions that are used by multiple modules
 in this subpackage.
 """
 
-from rmgpy.data.base import DatabaseError, LogicNode
-from rmgpy.reaction import Reaction, ReactionError
+from rmgpy.data.base import LogicNode
+from rmgpy.reaction import Reaction
 from rmgpy.molecule import Group
 from rmgpy.species import Species
-
-################################################################################
-
-class KineticsError(Exception):
-    """
-    An exception for problems with kinetics. Pass a string describing the problem.
-    """
-    pass
-
-class UndeterminableKineticsError(ReactionError):
-    """
-    An exception raised when attempts to estimate appropriate kinetic parameters
-    for a chemical reaction are unsuccessful.
-    """
-    def __init__(self, reaction, message=''):
-        new_message = 'Kinetics could not be determined. '+message
-        ReactionError.__init__(self,reaction,new_message)
+from rmgpy.exceptions import DatabaseError, KineticsError
 
 ################################################################################
