+++ rmgpy/data/kinetics/groups.py
@@ -38,12 +38,12 @@ import math
 import numpy
 from copy import deepcopy
 
-from rmgpy.data.base import Database, DatabaseError, Entry, Group, LogicNode, getAllCombinations, makeLogicNode
+from rmgpy.data.base import Database, Entry, Group, LogicNode, getAllCombinations, makeLogicNode
 
 from rmgpy.kinetics import Arrhenius, ArrheniusEP, KineticsData
 from rmgpy.species import Species
 from rmgpy.quantity import constants
-from .common import KineticsError, UndeterminableKineticsError
+from rmgpy.exceptions import KineticsError, UndeterminableKineticsError, DatabaseError
 
 ################################################################################
