+++ rmgpy/molecule/group.py
@@ -41,38 +41,7 @@ from .atomtype import atomTypes, allElements, nonSpecifics, getFeatures
 from .element import PeriodicSystem
 import rmgpy.molecule.molecule as mol
 from copy import deepcopy, copy
-
-################################################################################
-
-class ActionError(Exception):
-    """
-    An exception class for errors that occur while applying reaction recipe
-    actions. Pass a string describing the circumstances that caused the
-    exceptional behavior.
-    """
-    pass
-
-class ImplicitBenzeneError(Exception):
-    """
-    An exception class when encountering a group with too many implicit benzene
-    atoms. These groups are hard to create sample molecules and hard for users
-    to interpret. Pass a string describing the limitation.
-    """
-    pass
-
-class UnexpectedChargeError(Exception):
-    """
-    An exception class when encountering a group/molecule with unexpected charge
-    Curently in RMG, we never expect to see -2/+2 or greater magnitude charge,
-    we only except +1/-1 charges on nitrogen, oxygen, sulfur or specifically
-    carbon monoxide/monosulfide.
-
-    Attributes:
-    `graph` is the molecule or group object with the unexpected charge
-    """
-    def __init__(self, graph):
-        self.graph = graph
-
+from rmgpy.exceptions import ActionError, ImplicitBenzeneError, UnexpectedChargeError
 ################################################################################
 
 class GroupAtom(Vertex):
