+++ rmgpy/pdep/cse.pyx
@@ -41,15 +41,7 @@ from libc.math cimport exp, log, sqrt
 import rmgpy.constants as constants
 
 from rmgpy.pdep.me import generateFullMEMatrix
-
-################################################################################
-
-class ChemicallySignificantEigenvaluesError(Exception):
-    """
-    An exception raised when the reservoir state method is unsuccessful for
-    any reason. Pass a string describing the cause of the exceptional behavior.
-    """
-    pass
+from rmgpy.exceptions import ChemicallySignificantEigenvaluesError
 
 ################################################################################
